+++ frontend/src/components/guardians/HomeInformationSettings/HomeInformation.js
@@ -1,22 +1,22 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import Modal from 'react-modal';
 import { FiX, FiTrash2 } from "react-icons/fi";
+import HomeInformationForm from './HomeInformationForm';
 
 const HomeInformation = ({ isOpen, setIsOpen }) => {
-
-    const [data, setData] = useState({});
+    const [items, setItems] = useState([]);
 
     const closeModal = () => {
         setIsOpen(false);
     }
 
-    const inputData = (e) => {
-        const { name, value } = e.target;
-        setData(prevData => ({
-            ...prevData,
-            [name] : value
-        }))
+    const addItem = (item) => {
+        setItems(prevData => [...prevData, item])
     }
+
+    useEffect(() => {
+        console.log(items)
+    }, [items])
 
     return (
         <Modal
@@ -29,12 +29,15 @@ const HomeInformation = ({ isOpen, setIsOpen }) => {
             <button className='p-2 mr-2 rounded-[50px] border-2 border-black hover:scale-110'><FiTrash2 /></button>
             <button className='p-2 rounded-[50px] border-2 border-black hover:scale-110' onClick={closeModal}><FiX /></button>
         </div>
+        <div className='absolute top-16 left-5 right-5 text-2xl'>
+        {items.map((item, index) => (
+            <div className='items-center border-transparent bg-white shadow-[2px_4px_1px_#a5996e] rounded-[50px] my-6 py-4' 
+                key={index}>{item.content}
+            </div>
+        ))}
+        </div>
         <div className='absolute bottom-5 right-5 left-5 text-2xl flex justify-center'>
-            <input className='w-[350px] py-2 px-3 mr-3 rounded-[50px] bg-secondary border-2 border-transparent focus:border-white outline-none'
-                type='text'
-                onChange={inputData}
-            />
-            <button  className='p-2 rounded-[50px] bg-secondary border-2 border-transparent focus:border-white hover:scale-110'>추가</button>
+            <HomeInformationForm addItem={addItem}/>
         </div>
         </Modal>
     );
