+++ frontend/src/components/guardians/HomeInformationSettings/HomeInformationFolder.js
@@ -12,7 +12,8 @@ const HomeInformationFolder = ({ folder }) => {
         (folder && 
         <div className='flex mx-9'>
             <div className='h-44 w-44 bg-tertiay rounded-[70px] flex items-center justify-center shadow-[2px_4px_1px_#a5996e] cursor-pointer'
-            onClick={openModal}></div>
+                onClick={openModal}>
+            </div>
             <HomeInformation isOpen={isOpen} setIsOpen={setIsOpen}/>
         </div>
         )
